change identifiers for public functions
In keeping with the conventions and guidelines outlined here:
https://github.com/bbatsov/emacs-lisp-style-guide

While Elisp doesn't have private or public functions per se, some Elisp
developers find it useful to encode functions that are intended to be private
with namespace--function-name while public functions get
namespace-function-name.

Additionally, the "item-" was removed from the public function since it felt
somewhat redundant. "-for" was added as appendage.

Lastly a sharp-quote was used in the call to (apply ...)